Fix clang-tidy warnings

+++ signal/sc_signal.c
@@ -521,6 +521,6 @@ void sc_signal_log(int fd, char *buf, size_t sz, char *fmt, ...)
 	wr = sc_signal_vsnprintf(buf, sz, fmt, args);
 	va_end(args);
 
-	rc = write(fd, buf, (size_t) wr);
+	rc = (int) write(fd, buf, (size_t) wr);
 	(void) rc;
 }
